# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Health Diet Online Search",
      "description": "A community for Health experts and users"
    },
    "host": "localhost:3000",
    "basePath": "/api/",
...
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
      "/queries/add": {
        "post": {
          "tags": ["Queries"],
          "summary": "add new Query",
          "description": "",
          "operationId": "addQuery",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Query object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Query"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Query created successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/get/{id}": {
        "get": {
          "tags": ["Queries"],
          "summary": "Get Query by id",
          "description": "",
          "operationId": "getQueryById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Query to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Query"
              }
            },
            "404": {
              "description": "This Query doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/interested": {
        "get": {
          "tags": ["Queries"],
          "summary": "Get Queries interested by current expert",
          "description": "",
          "operationId": "getQuery",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Query"
              }
            },
            "404": {
              "description": "no queries found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/get/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Queries"],
          "summary": "Get total Queries",
          "description": "",
          "operationId": "getQueries",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 posts per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Query"
              }
            },
            "404": {
              "description": "No Queries found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/get/myqueries/{pagesize}/{pagenum}": {
        "get": {
          "tags": ["Queries"],
          "summary": "Get total Query of the logedin user",
          "description": "",
          "operationId": "getQueryOfLogedinUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "pagesize",
              "description": "size of the page like 5 for 5 Query per page",
              "required": true,
              "type": "string"
            },
            {
              "in": "path",
              "name": "pagenum",
              "description": "page number",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "schema": {
                "$ref": "#/definitions/Query"
              }
            },
            "404": {
              "description": "No Query found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/update/{id}": {
        "put": {
          "tags": ["Queries"],
          "summary": "update Query",
          "description": "",
          "operationId": "updateQuery",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Query to be updated",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Query object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Query"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This Query doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/queries/delete/{id}": {
        "delete": {
          "tags": ["Queries"],
          "summary": "delete Query by id",
          "description": "",
          "operationId": "deleteQueryById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of Query to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Query deleted"
            },
            "401": {
              "description": "access denied."
            },
            "404": {
              "description": "The Query with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/solutions/add": {
        "post": {
          "tags": ["Solutions"],
          "summary": "add new Solution",
          "description": "",
          "operationId": "addSolution",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Solution object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Solution"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Solution updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/solutions/update/{id}": {
        "put": {
          "tags": ["Solutions"],
          "summary": "Update Solution",
          "description": "",
          "operationId": "UpdateSolution",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Solution object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Solution"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Solution updated successfully"
            },
            "400": {
              "description": "Invalid user"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/solutions/delete/{id}": {
        "delete": {
          "tags": ["Solutions"],
          "summary": "delete solution by id",
          "description": "",
          "operationId": "deleteSolutionById",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "id of solution to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }

          ],
          "responses": {
            "200": {
              "description": "Solution deleted"
            },
            "404": {
              "description": "The Solution with the given ID was not found."
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
...
        "Query": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "data": {
              "type": "string"
            },
            "tags": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "Solution": {
          "type": "object",
          "properties": {
            "solution": {
              "type": "string"
            },
            "queryId" : {
              "type" : "string"
            }
          }
        },
...
      } 
}